+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # adrez-mapper Revisions
 
-
+2021-05-17 (1.6.4)
+- add: added force to mode so empty record can be handled
 
 2021-05-15 (1.6.3)
 - fix: _mode makes a record forced to store